Implements yml and descriptionService

+++ src/main/java/motolibrary/service/DescriptionService.java
@@ -4,6 +4,6 @@ import motolibrary.model.Languages;
 
 public interface DescriptionService {
 
-    String getDescriptionById(Languages lang, int id);
+    String getDescriptionById(Languages lang, long id);
 
 }
